More refactoring, made more stuff internal: - #1581
