+++ src/Microsoft.AspNet.Mvc.TagHelpers/Internal/TagHelperContextExtensions.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNet.Mvc.TagHelpers.Internal;
 using Microsoft.AspNet.Razor.Runtime.TagHelpers;
 using Microsoft.Framework.Logging;
 
-namespace Microsoft.AspNet.Mvc.TagHelpers
+namespace Microsoft.AspNet.Mvc.TagHelpers.Internal
 {
     /// <summary>
     /// Utility related extensions for <see cref="TagHelperContext"/>.
